// frontend/src/app/pages/admin-panel-page/models/offer-request-dto.ts
import { GovernmentIdDTO } from "./government-id-dto";

export interface OfferRequestDTO {
    id: string;
    loanValue: number;
    numberOfInstallments: number;
    percentage: number;
    monthlyInstallment: number;
    status: string;
    email: string;
    dateOfInquiry: Date;
    dateOfApplication: Date;
    governmentId: GovernmentIdDTO;
}